without registration to database

+++ src/main/java/com/iba/YourReview/repository/MovieRepo.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface MovieRepo extends CrudRepository<Movie, Integer> {
+public interface MovieRepo extends CrudRepository<Movie, Long> {
 
     Movie findByTitle(String title);
